+++ src/js/mixin/internal/animate-fade.js
@@ -11,6 +11,7 @@ import {
     Transition,
 } from 'uikit-util';
 import { getRows } from '../../core/margin';
+import { awaitFrame } from './animate-slide.js';
 
 const clsLeave = 'uk-transition-leave';
 const clsEnter = 'uk-transition-enter';
@@ -117,13 +118,7 @@ function waitTransitionend(target) {
 }
 
 function getTransitionNodes(target) {
-    return getRows(children(target))
-        .flat()
-        .filter((node) => isVisible(node));
-}
-
-function awaitFrame() {
-    return new Promise((resolve) => requestAnimationFrame(resolve));
+    return getRows(children(target)).flat().filter(isVisible);
 }
 
 function awaitTimeout(timeout) {