+++ src/js/util/svg.js
@@ -1,16 +1,8 @@
 import { $$ } from './dom';
+import { isVisible } from './filter.js';
 
 export function getMaxPathLength(el) {
-    return Math.ceil(
-        Math.max(
-            0,
-            ...$$('[stroke]', el).map((stroke) => {
-                try {
-                    return stroke.getTotalLength();
-                } catch (e) {
-                    return 0;
-                }
-            }),
-        ),
-    );
+    return isVisible(el)
+        ? Math.ceil(Math.max(0, ...$$('[stroke]', el).map((stroke) => stroke.getTotalLength())))
+        : 0;
 }